+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `settings.misc.filter` and `ignore_filters` argument.
+
 ## [0.4.0] - 2024-06-17
 
 ### Deprecated